Handle DB write conflict in the workflow service
For a workflow with parallel branches, it is possible for the workflow service to handle action execution completion for more than one tasks. When this happens, it is likely that a DB write conflict can occur. To resolve, the workflow service needs to catch the DB exception and then retry the logic to handle action execution completion. The logic is also divided into multiple atomic parts.